Moved and renamed user folder (now userDefined) outside of main LoomSE folder
