+++ app-src/LoomSE/Models/Scene.js
@@ -1,6 +1,6 @@
 import view from '../view';
 import storyBehaviour from '../../constants/storyBehaviour';
-import * as userDefinedModules from '../user/userModules';
+import * as userDefinedModules from '../../userDefined/userModules';
 
 import { Events, Subtitles, Video } from '../Components';
 import { parseFile } from '../tools/fileParsers';
